fix message splicing
