+++ core/llm/messages.ts
@@ -1,0 +1,68 @@
+import { ChatMessage } from "..";
+
+export function messageHasToolCalls(msg: ChatMessage): boolean {
+  return msg.role === "assistant" && !!msg.toolCalls;
+}
+
+export function flattenMessages(msgs: ChatMessage[]): ChatMessage[] {
+  const flattened: ChatMessage[] = [];
+
+  for (let i = 0; i < msgs.length; i++) {
+    const msg = msgs[i];
+
+    if (
+      flattened.length > 0 &&
+      flattened[flattened.length - 1].role === msg.role &&
+      !messageHasToolCalls(msg) &&
+      !messageHasToolCalls(flattened[flattened.length - 1])
+    ) {
+      flattened[flattened.length - 1].content += `\n\n${msg.content || ""}`;
+    } else {
+      flattened.push(msg);
+    }
+  }
+
+  return flattened;
+}
+
+export function messageIsEmpty(message: ChatMessage): boolean {
+  if (typeof message.content === "string") {
+    return message.content.trim() === "";
+  }
+  if (Array.isArray(message.content)) {
+    return message.content.every(
+      (item) => item.type === "text" && item.text?.trim() === "",
+    );
+  }
+  return false;
+}
+
+export function addSpaceToAnyEmptyMessages(
+  messages: ChatMessage[],
+): ChatMessage[] {
+  return messages.map((message) => {
+    if (messageIsEmpty(message)) {
+      message.content = " ";
+    }
+    return message;
+  });
+}
+
+export function chatMessageIsEmpty(message: ChatMessage): boolean {
+  switch (message.role) {
+    case "system":
+    case "user":
+      return (
+        typeof message.content === "string" && message.content.trim() === ""
+      );
+    case "assistant":
+      return (
+        typeof message.content === "string" &&
+        message.content.trim() === "" &&
+        !message.toolCalls
+      );
+    case "thinking":
+    case "tool":
+      return false;
+  }
+}
